fix(Indexer): Workaround "Deadlock of Hell"
There is an issue I've had during development where a shared ILogger instance in Parallel.ForEach would cause ForEach to never complete. I lovingly coined this the "Deadlock of Hell".

I've accidentally brought it back with 09a83f6 because the Parallel.ForEach uses the shared context logger, this fixes that.