fix: Ensure Indexer runs before Patcher by splitting Host.Run
Since we pull in Patcher in Main, we run the Patcher before the Indexer (which actually hydrates the database with data), so that should fail on first patch.

Instead, we call host.StartAsync to initialize all the Services, which allows us to then operate on the state of all those services.

This technically starts Live, but we'd have to repeat the config hack from before to fix that I think.